# Drop problem sizes for some slow running tests