--- a/test/library/packages/LinearAlgebra/performance/me/me-perf.execopts
+++ b/test/library/packages/LinearAlgebra/performance/me/me-perf.execopts
@@ -1,1 +1,1 @@
---correctness=true
+--correctness=true --iters=1 --iterLimit=1