--- a/test/library/standard/Random/RandomStreamInterface/choiceTestUtils.chpl
+++ b/test/library/standard/Random/RandomStreamInterface/choiceTestUtils.chpl
@@ -2,7 +2,7 @@ use Random, Map;
 
 config const debug = false;
 
-proc test(stream, X, size:?sizeType=none, replace=true, prob:?probType=none, trials=10000) throws {
+proc test(stream, X, size:?sizeType=none, replace=true, prob:?probType=none, trials=1000) throws {
   var counts;
   if isArrayValue(X) {
     counts = new map(X.eltType, int);