Decrease some problem sizes for:
 - me-perf: 240s -> 1s
 - seeking: 120s -> 1s (for all 4 execopts on NFS)
 - choiceTest{Arr,Dom,Rng}: 200s -> 20s (for all three tests)

Part of Cray/chapel-private#3211